Start GUI V3
After completing the second version of the start GUI I decided to add a little bit more onto the program before moving on. This involved disabling the buttons before the user enters an amount of money, and then only making the buttons available that fit the amount of money that the user put in. Otherwise the program was quite similar, also with error handling that told the user if they entered an invalid input.